refactor: migrate page transition builder to widgets
