+++ packages/flutter/lib/src/material/page.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/widgets.dart';
 
 import 'page_transitions_theme.dart';
 import 'theme.dart';
